requirements.txt
pandas==2.1.4
requests==2.31.0
ephem==4.1.5
plotly==5.18.0
python-dotenv==1.0.0
numpy==1.26.2
scipy==1.11.4 